requirements.txt
Flask
flask-smorest
marshmallow
python-dotenv
pymongo
Werkzeug 
PyJWT